Add Moodle 4.5 support (see MDL-82427)
